third_party/WebKit/LayoutTests/inspector/sources/sources-panel-extension-names-expected.txt

The test verifies that extension names are resolved properly in navigator view.


Running: testAddExecutionContextBeforeFile
  Dumping ScriptsNavigator contentScripts tab:
    FibExtension
